refactor(settings): extract domain workflows
