+++ src/components/settings/SettingsDialog.vue
@@ -4,7 +4,7 @@ import { computed } from 'vue'
 import { useI18n } from '@open-pencil/vue'
 import { IS_TAURI } from '@open-pencil/core/constants'
 
-import { useAIChat } from '@/app/ai/chat/use'
+import { browserCredentialsRemembered, setRememberCredentials } from '@/app/ai/chat/storage'
 import { appCredentialServices } from '@/app/settings/credentials/app'
 import { settingsDialogOpen, settingsDialogSection } from '@/app/settings/dialog'
 import DiagnosticsSettingsPanel from '@/components/settings/diagnostics/DiagnosticsSettingsPanel.vue'
@@ -20,7 +20,6 @@ import AppSwitch from '@/components/ui/AppSwitch.vue'
 import { AppDialogFooter, AppDialogHeader, AppDialogRoot } from '@/components/ui/dialog'
 
 const { credentials, settings, common } = useI18n()
-const { browserCredentialsRemembered, setRememberCredentials } = useAIChat()
 function onOpenChange(open: boolean): void {
   settingsDialogOpen.value = open
 }
