+++ src/components/settings/mcp/MCPSettingsPanel.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
-import type { ToolEffect } from '@open-pencil/mcp/tools'
-import { computed, onMounted, ref } from 'vue'
+import { useMCPSettings } from '@/app/automation/mcp/settings/use'
+
+import { computed } from 'vue'
 import { useI18n } from '@open-pencil/vue'
 
 import {
@@ -11,7 +12,7 @@ import {
   setMCPToolCategoryEnabled,
   setMCPToolEnabled
 } from '@/app/automation/mcp/preferences'
-import { mcpRuntime, refreshMCPRuntime, restartMCPRuntime } from '@/app/automation/mcp/runtime'
+import { mcpRuntime } from '@/app/automation/mcp/runtime'
 import { isTauri } from '@/app/tauri/env'
 import AppInput from '@/components/ui/AppInput.vue'
 import AppSwitch from '@/components/ui/AppSwitch.vue'
@@ -33,52 +34,18 @@ const statusMessage = computed(() => {
       return automation.value.statusIdle
   }
 })
-const toolSearch = ref('')
-const disabledToolNames = computed(() => new Set(disabledMCPTools.value))
-function categoryStatus(effect: ToolEffect) {
-  const tools = configurableMCPTools.value.filter((tool) => tool.effect === effect)
-  const enabled = tools.filter((tool) => !disabledToolNames.value.has(tool.name)).length
-  return {
-    enabled: enabled > 0,
-    state: enabled > 0 && enabled < tools.length ? ('mixed' as const) : ('idle' as const)
-  }
-}
-const inspectionToolsStatus = computed(() => categoryStatus('read'))
-const modificationToolsStatus = computed(() => categoryStatus('write'))
-const enabledToolCount = computed(
-  () => configurableMCPTools.value.filter((tool) => !disabledToolNames.value.has(tool.name)).length
-)
-const visibleTools = computed(() => {
-  const query = toolSearch.value.trim().toLowerCase()
-  if (!query) return configurableMCPTools.value
-  return configurableMCPTools.value.filter(
-    (tool) =>
-      tool.name.toLowerCase().includes(query) || tool.description.toLowerCase().includes(query)
-  )
-})
 
-onMounted(() => {
-  void refreshMCPRuntime()
-})
-
-function restart(): void {
-  void restartMCPRuntime()
-}
-
-async function chooseRootDirectory(): Promise<void> {
-  if (!isTauri()) return
-  const { open } = await import('@tauri-apps/plugin-dialog')
-  const directory = await open({ directory: true, multiple: false })
-  if (typeof directory === 'string') mcpRootDirectory.value = directory
-}
-
-function isToolEnabled(name: string): boolean {
-  return !disabledToolNames.value.has(name)
-}
-
-function enableAllTools(): void {
-  disabledMCPTools.value = []
-}
+const {
+  toolSearch,
+  inspectionToolsStatus,
+  modificationToolsStatus,
+  enabledToolCount,
+  visibleTools,
+  restart,
+  chooseRootDirectory,
+  isToolEnabled,
+  enableAllTools
+} = useMCPSettings()
 </script>
 
 <template>
